+++ src/core/light/PointLight2.ts
@@ -2,21 +2,31 @@ import {Color, ColorRepresentation, PointLight, PointLightShadow, Vector2, Vecto
 import {ILight} from './ILight'
 import {iLightCommons} from '../object/iLightCommons'
 import {IObject3D} from '../IObject'
-import {uiColor, uiFolderContainer, uiNumber, UiObjectConfig, uiSlider, uiToggle, uiVector} from 'uiconfig.js'
+import {generateUiConfig, uiColor, uiNumber, uiSlider, uiToggle, uiVector} from 'uiconfig.js'
 import {onChange3} from 'ts-browser-helpers'
 import {bindToValue} from '../../three'
+import {objectActionsUiConfig} from '../object/IObjectUi.ts'
 
 /**
  * Extension of three.js PointLight with additional properties for serialization and UI
+ * A point light emits light in all directions from a single point in space.
  *
  * Note - gltf serialization is handled by {@link GLTFLightExtrasExtension}
+ *
+ * @category Lights
  */
-@uiFolderContainer('Point Light')
 export class PointLight2 extends PointLight implements ILight<PointLightShadow> {
     assetType = 'light' as const
     setDirty = iLightCommons.setDirty
     refreshUi = iLightCommons.refreshUi
-    declare uiConfig: UiObjectConfig
+    uiConfig = {
+        type: 'folder',
+        label: 'Point Light',
+        children: [
+            ...generateUiConfig(this),
+            ...objectActionsUiConfig.call(this),
+        ],
+    }
     readonly isPointLight2 = true
 
     @uiToggle('Enabled')