+++ src/core/light/RectAreaLight2.ts
@@ -2,15 +2,30 @@ import {Color, ColorRepresentation, RectAreaLight} from 'three'
 import {ILight} from './ILight'
 import {iLightCommons} from '../object/iLightCommons'
 import {IObject3D} from '../IObject'
-import {uiColor, uiFolderContainer, uiNumber, UiObjectConfig, uiSlider, uiToggle} from 'uiconfig.js'
+import {generateUiConfig, uiColor, uiNumber, uiSlider, uiToggle} from 'uiconfig.js'
 import {onChange3} from 'ts-browser-helpers'
+import {objectActionsUiConfig} from '../object/IObjectUi.ts'
 
-@uiFolderContainer('RectArea Light')
+/**
+ * Extension of three.js RectAreaLight with additional properties for serialization and UI
+ * A rectangular area light emits light from a rectangular plane, commonly used to simulate window lights or panels.
+ *
+ * Note - gltf serialization is handled by {@link GLTFLightExtrasExtension}
+ *
+ * @category Lights
+ */
 export class RectAreaLight2 extends RectAreaLight implements ILight<undefined> {
     assetType = 'light' as const
     setDirty = iLightCommons.setDirty
     refreshUi = iLightCommons.refreshUi
-    uiConfig: UiObjectConfig
+    uiConfig = {
+        type: 'folder',
+        label: 'RectArea Light',
+        children: [
+            ...generateUiConfig(this),
+            ...objectActionsUiConfig.call(this),
+        ],
+    }
     readonly isRectAreaLight2 = true
 
     @uiToggle('Enabled')